backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
sqlite3
faiss-cpu
numpy
rank-bm25
sentence-transformers
scikit-learn
pyyaml
sentry-sdk
streamlit
pandas
plotly
jinja2
